+++ src/main/java/com/server/crews/global/config/DatabaseInitializer.java
@@ -1,6 +1,5 @@
 package com.server.crews.global.config;
 
-import com.redis.lettucemod.api.StatefulRedisModulesConnection;
 import com.server.crews.applicant.domain.Application;
 import com.server.crews.applicant.domain.NarrativeAnswer;
 import com.server.crews.applicant.domain.SelectiveAnswer;
@@ -15,12 +14,11 @@ import com.server.crews.recruitment.domain.Recruitment;
 import com.server.crews.recruitment.domain.Section;
 import com.server.crews.recruitment.domain.SelectiveQuestion;
 import com.server.crews.recruitment.repository.RecruitmentRepository;
-import com.server.crews.recruitment.repository.RecruitmentSearchKeywordRepository;
-import com.server.crews.recruitment.service.RecruitmentSearchService;
+import com.server.crews.recruitment.service.SimpleRedisRecruitmentSearchService;
+import com.server.crews.recruitment.service.RedisStackRecruitmentSearchService;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.apache.commons.pool2.impl.GenericObjectPool;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.context.annotation.Profile;
@@ -38,8 +36,8 @@ public class DatabaseInitializer implements ApplicationRunner {
     private final RecruitmentRepository recruitmentRepository;
     private final ApplicantRepository applicantRepository;
     private final ApplicationRepository applicationRepository;
-    private final RecruitmentSearchKeywordRepository recruitmentSearchKeywordRepository;
-    private final RecruitmentSearchService recruitmentSearchService;
+    private final SimpleRedisRecruitmentSearchService simpleRedisRecruitmentSearchService;
+    private final RedisStackRecruitmentSearchService redisStackRecruitmentSearchService;
 
     @Override
     public void run(ApplicationArguments args) {
@@ -73,9 +71,9 @@ public class DatabaseInitializer implements ApplicationRunner {
                 List.of(commonSection, backendSection, frontendSection));
         recruitment.close();
         recruitmentRepository.save(recruitment);
-        recruitmentSearchKeywordRepository.saveRecruitment(recruitment);
-        recruitmentSearchService.createIndex();
-        recruitmentSearchService.saveRecruitment(recruitment);
+        simpleRedisRecruitmentSearchService.saveRecruitment(recruitment);
+        redisStackRecruitmentSearchService.createIndex();
+        redisStackRecruitmentSearchService.saveRecruitment(recruitment);
 
         Applicant kh = new Applicant("kh@google.com", passwordEncoder.encode("test-password"));
         Applicant lkh = new Applicant("lkh@google.com", passwordEncoder.encode("test-password"));
